refactor(routes): split routes/settings.py into sub-router package
routes/settings.py had grown to 1002 lines and nine routes with mixed
instance CRUD, account password change, and test-connection plumbing.
Convert it into a routes/settings/ package with four focused modules:

* __init__.py composes a single settings_router the app mounts unchanged
* _helpers.py owns rendering, client construction, the connection probe,
  the validation helpers, and the _render_settings_page composition the
  page and account routes both delegate to
* page.py owns GET /settings
* account.py owns POST /settings/account/password
* instances.py owns the seven /settings/instances/* routes

Every helper now lives behind a stable public name (no leading
underscore on cross-module callers) so sub-modules do not reach into a
sibling's private API. No route paths, payloads, or template
selections change; the full non-integration suite (1323 tests) passes.
